Map alterations, platform trigger boxes
Added trigger boxes to all platforms and added functionality to change gravity in any trigger box from UE4 editor. Added surface below water level for the ball to land on and made visuble water surface have no impact mechanic. Ball passes through and sits on subsurface as if it was in water or liquid.