Fixed switch impact issue and added assist lift
Fixed the switches being hit several times when throw the ball at them by added a timer. Timer is set to allow a hit every second.
A lift was added to the opposite side of the switches allowing the player to return to the starting landing after falling off. Player can still stand under the lift. It will push the player through it and onto the lift. Will fix later. Spotlight issues persist.