Began work on triggers to play dialogue
Added an audio component to the player class. Created a trigger box to play a specific sound cue through the player's audio component. Imported a test .wav file and created a test sound cue to play when the trigger box is entered by the player. Still need to make it so the sound cue only plays once and does not get interupted by other sound cues.